+++ lib/services/window_service.dart
@@ -1,9 +1,9 @@
 import 'dart:ui';
 
+import 'package:lotti/database/settings_db.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/sync/inbox/inbox_service.dart';
 import 'package:lotti/sync/outbox/outbox_service.dart';
-import 'package:lotti/sync/secure_storage.dart';
 import 'package:window_manager/window_manager.dart';
 
 class WindowService implements WindowListener {
@@ -11,8 +11,8 @@ class WindowService implements WindowListener {
     windowManager.addListener(this);
   }
 
-  final sizeKey = 'sizeKey';
-  final offsetKey = 'offsetKey';
+  final sizeKey = 'WINDOW_SIZE';
+  final offsetKey = 'WINDOW_OFFSET';
 
   Future<void> restore() async {
     await restoreSize();
@@ -20,7 +20,7 @@ class WindowService implements WindowListener {
   }
 
   Future<void> restoreSize() async {
-    final sizeString = await getIt<SecureStorage>().readValue(sizeKey);
+    final sizeString = await getIt<SettingsDb>().itemByKey(sizeKey);
     final values = sizeString?.split(',').map(double.parse).toList();
     final width = values?.first;
     final height = values?.last;
@@ -30,7 +30,7 @@ class WindowService implements WindowListener {
   }
 
   Future<void> restoreOffset() async {
-    final offsetString = await getIt<SecureStorage>().readValue(offsetKey);
+    final offsetString = await getIt<SettingsDb>().itemByKey(offsetKey);
     final values = offsetString?.split(',').map(double.parse).toList();
     final dx = values?.first;
     final dy = values?.last;
@@ -63,19 +63,27 @@ class WindowService implements WindowListener {
   @override
   void onWindowMinimize() {}
 
-  @override
-  Future<void> onWindowMove() async {
+  Future<void> _onMoved() async {
     final offset = await windowManager.getPosition();
-    await getIt<SecureStorage>()
-        .writeValue(offsetKey, '${offset.dx},${offset.dy}');
+    await getIt<SettingsDb>().saveSettingsItem(
+      offsetKey,
+      '${offset.dx},${offset.dy}',
+    );
+  }
+
+  Future<void> _onResized() async {
+    final size = await windowManager.getSize();
+    await getIt<SettingsDb>().saveSettingsItem(
+      sizeKey,
+      '${size.width},${size.height}',
+    );
   }
 
   @override
-  Future<void> onWindowResize() async {
-    final size = await windowManager.getSize();
-    await getIt<SecureStorage>()
-        .writeValue(sizeKey, '${size.width},${size.height}');
-  }
+  Future<void> onWindowMove() async {}
+
+  @override
+  Future<void> onWindowResize() async {}
 
   @override
   void onWindowRestore() {}
@@ -87,8 +95,12 @@ class WindowService implements WindowListener {
   void onWindowClose() {}
 
   @override
-  void onWindowMoved() {}
+  void onWindowMoved() {
+    _onMoved();
+  }
 
   @override
-  void onWindowResized() {}
+  void onWindowResized() {
+    _onResized();
+  }
 }