+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed:
 - Locking issue in sync
 
+### Added:
+- Settings DB
+
+### Changed:
+- Window manager persistence moved to settings database
+
 ## [0.8.255] - 2023-02-02
 ### Changed:
 - Upgrade dependencies